refs #81983: Refactor interactive region layers into one and make missing regions properly update the map
