+++ apps/src/types/types.ts
@@ -19,7 +19,7 @@ export type DatasetKey = 'cmip5' | 'cmip6' | 'humidex';
 /**
  * Represents valid emission scenario keys.
  */
-export type EmissionScenarioKey = 'low' | 'medium' | 'high';
+export type EmissionScenarioKey = 'low' | 'medium' | 'high' | 'very-high';
 
 /**
  * Represents a field with multilingual support.
@@ -197,6 +197,7 @@ export interface MapState {
 	pane: string;
 	dataValue: string;
 	mapColor: string;
+	legendData: WMSLegendData;
 	opacity: {
 		mapData: number;
 		labels: number;
@@ -235,6 +236,7 @@ export interface WMSLegendEntry {
 	label: string;
 	color: string;
 	opacity: string;
+	quantity: number;
 }
 
 /**
