+++ fw-child/resources/js/map.js
@@ -957,6 +957,7 @@
 
           options.grid.highlighted = shape_id;
 
+          console.log('aaaa', shape_id, style_obj);
           for (let key in options.maps) {
             options.maps[key].layers.grid.setFeatureStyle(shape_id, style_obj);
           }
